- Added cacheImage - MovieApi: remove getImage function - ImageProvider: added singleton for get the image and use the cache
