+++ BAZProject/DetailMovieViper/CastViper/Presenter/DetailMovieCastPresenter.swift
@@ -42,7 +42,7 @@ class DetailMovieCastPresenter: DetailMovieCastPresenterProtocol {
     /// - Parameter completion: Escaping closure that escapes a UIImage or a nil
     /// - Returns: escaping closure with the UIImage type, if the parse fails, can return nil
     func getCastImage(index: Int, completion: @escaping (UIImage?) -> Void) {
-        movieApi.getImage(for: cast[index].profile_path ?? "") { castImage in
+        ImageProvider.shared.getImage(for: cast[index].profile_path ?? "") { castImage in
             completion(castImage)
         }
     }
