+++ BAZProject/DetailMovieViper/RecommendationViper/Presenter/DetailMovieRecommendationPresenter.swift
@@ -42,7 +42,7 @@ class DetailMovieRecommendationPresenter: DetailMovieRecommendationPresenterProt
     /// - Parameter completion: Escaping closure that escapes a UIImage or a nil
     /// - Returns: escaping closure with the UIImage type, if the parse fails, can return nil
     func getRecommendationImage(index: Int, completion: @escaping (UIImage?) -> Void) {
-        movieApi.getImage(for: recommendation[index].poster_path ?? "") { recommendationImage in
+        ImageProvider.shared.getImage(for: recommendation[index].poster_path ?? "") { recommendationImage in
             completion(recommendationImage)
         }
     }
