+++ BAZProject/DetailMovieViper/SimilarViper/Presenter/DetailMovieSimilarPresenter.swift
@@ -42,7 +42,7 @@ class DetailMovieSimilarPresenter: DetailMovieSimilarPresenterProtocol {
     /// - Parameter completion: Escaping closure that escapes a UIImage or a nil
     /// - Returns: escaping closure with the UIImage type, if the parse fails, can return nil
     func getSimilarImage(index: Int, completion: @escaping (UIImage?) -> Void) {
-        movieApi.getImage(for: similar[index].poster_path ?? "") { similarImage in
+        ImageProvider.shared.getImage(for: similar[index].poster_path ?? "") { similarImage in
             completion(similarImage)
         }
     }
