+++ BAZProject/HomeMovieViper/View/Base.lproj/HomeMoviesView.storyboard
@@ -72,7 +72,7 @@
                         </constraints>
                     </view>
                     <navigationItem key="navigationItem" title="Movies" id="WhG-gT-S6I">
-                        <barButtonItem key="rightBarButtonItem" style="plain" systemItem="search" id="6Jw-jj-qME">
+                        <barButtonItem key="rightBarButtonItem" style="done" systemItem="search" id="6Jw-jj-qME">
                             <color key="tintColor" systemColor="labelColor"/>
                             <connections>
                                 <action selector="didSelectSearchButton:" destination="xsT-Cm-LUD" id="wdW-1t-MPB"/>
