+++ BAZProject/MovieAPI.swift
@@ -152,22 +152,4 @@ class MovieAPI {
             }
         }
     }
-    
-    /// Get an image from the API movies and convert the url to an UIImage
-    ///
-    /// - Parameter url: Image url
-    /// - Parameter completion: Escaping closure that escapes a UIImage or a nil
-    /// - Returns: Escaping closure with the uiImage type, if the parse fails, can return nil
-    func getImage(for urlIdentifierImage: String, completion: @escaping (UIImage?) -> Void) {
-        let urlString = "https://image.tmdb.org/t/p/w500\(urlIdentifierImage)"
-        DispatchQueue.global(qos: .background).async {
-            if let url = URL(string: urlString),
-               let data = try? Data(contentsOf: url),
-               let image: UIImage = UIImage(data: data) {
-                 completion(image)
-            } else {
-                completion(nil)
-            }
-        }
-    }
 }
